staging-causality (explainable reasoning as a separate module) (#441)
* removed explanation code (#411)

* add `preempt` and its tests (#414)

* added preempt and its tests

* refactored test to use new code

* lint

* added uniform proposal (#416)

* add consequent_differs and a test thereof (#418)

* added preempt and its tests

* refactored test to use new code

* lint

* added consequent_differs and a test thereof

---------


* add Preemptions and a test thereof (#420)

* added preempt and its tests

* refactored test to use new code

* lint

* added Preemptions and a test thereof

* added `undo_split` and a test thereof (#422)

* import

---------


* added SplitSubsets and a test thereof (#424)

* added preempt and its tests

* refactored test to use new code

* lint

* added Preemptions and a test thereof

* added `undo_split` and a test thereof

* added SplitSubsets and a test thereof

* added `undo_split` and a test thereof (#422)

* import

---------


* added `random intervention` and a test thereof (#442)

* add `SearchForExplanation` (#445)

* added `random intervention` and a test thereof

* added `SearchForExplanation` and a test thereof

---------


* explainable documentation update (#457)

* docstrings update WIP

* small typo

* Remove explainable references from chirho.counterfactual (#458)

* implement `soft_eq` and `soft_neq`  (#472)

* add soft_neq and a few tests in test_ops

* sof eq WIP

* conversion to soft_eq

* defaulting soft_eq scale to .1

* Tavares conditions have landed

* soft_neq with tests

* added docstring for soft_neq

* refactor soft_eq and soft_neq

* tests

* remove failing tests

* comment

* sign

* move soft_neq to internals for now

* nit

* remove comment

---------


* Add explainable module to Sphinx build (#481)

* Add explainable module to sphinx build

* Add explainable module to sphinx build

* Reorganize code in chirho.explainable (#485)

* Add explainable module to sphinx build

* Add explainable module to sphinx build

* reorganize codebase

* remove empty file

* rename alternatives

* reorganize test files

* sphinx

* remove test_defaults

* implement soft_neq and use downstream in the explainable module (#490)

* that's it

* lint

* lint

* preemptions type as required by runner lint

* persistent lint typing error, using Any

* reverting to Preemptions

* add revised actual causality notebook (#446)

* added `random intervention` and a test thereof

* added `SearchForExplanation` and a test thereof

* added ac notebook with a test

* small cleanup

* that's it

* rerun actual causality and added it to rst

* actual causality nb WIP

* lint

* proofreading

* lint

* preemptions type as required by runner lint

* persistent lint typing error, using Any

* reverting to Preemptions

* lint, add actual_causality to notebook tests again

* removed redundant alternatives.py

* remove redundant split_subset.py

* fixing docstrings

* format & lint

* pulled explanation from ru-propagate-...

* docstrings in explanation.py

* typo in ac nb

---------


* allow for empty witness in SearchForCause (#491)

* allow for empty witness in SearchForCause

* removed mwc_empty from a test

* allow empty ant and con

* Implement `InferConstraints` (#493)

* experiment done

* cleanup, docstring

* eliminate args from ExtractSupport, move to components

* lint

---------
