+++ docs/source/counterfactual.rst
@@ -31,10 +31,6 @@ Handlers
    :members:
    :undoc-members:
 
-.. automodule:: chirho.counterfactual.handlers.explanation
-   :members:
-   :undoc-members:
-
 Internals
 ---------
