+++ docs/source/explainable.rst
@@ -1,0 +1,43 @@
+Explainable
+===========
+
+.. automodule:: chirho.explainable
+   :members:
+   :undoc-members:
+
+Operations
+----------
+
+.. automodule:: chirho.explainable.ops
+   :members:
+   :undoc-members:
+
+Handlers
+--------
+
+.. automodule:: chirho.explainable.handlers
+   :members:
+   :undoc-members:
+
+.. automodule:: chirho.explainable.handlers.components
+   :members:
+   :undoc-members:
+
+.. automodule:: chirho.explainable.handlers.explanation
+   :members:
+   :undoc-members:
+
+.. automodule:: chirho.explainable.handlers.preemptions
+   :members:
+   :undoc-members:
+
+Internals
+---------
+
+.. automodule:: chirho.explainable.internals
+   :members:
+   :undoc-members:
+
+.. automodule:: chirho.explainable.internals.defaults
+   :members:
+   :undoc-members:
