+++ docs/source/index.rst
@@ -31,6 +31,7 @@ Table of Contents
    sdid
    dr_learner
    dynamical_intro
+   actual_causality
 
 .. toctree::
    :maxdepth: 2
@@ -41,6 +42,7 @@ Table of Contents
    observational
    indexed
    dynamical
+   explainable
 
 .. toctree::
    :maxdepth: 2
